refactor : 주행 조건 전략 패턴으로 대치, 기능별 패키지 상세화
Refactor
- 주행 조건을 RacingCar 클래스에서 빼내 전략 패턴으로 구현, 도메인 구체화
- 기능별 패키지 상세화 car, strategy, view 등
