+++ src/main/java/edu/nextstep/racingcar/strategy/RacingStrategy.java
@@ -1,0 +1,7 @@
+package edu.nextstep.racingcar.strategy;
+
+public interface RacingStrategy {
+
+    public abstract boolean canMove();
+
+}
